small unimportant changes

+++ cosas/include/cosas/filter.h
@@ -91,15 +91,15 @@ private:
 class CtrlDamper {
 
 public:
-  CtrlDamper(std::array<uint8_t, N_KNOBS> lo, std::array<uint8_t, N_KNOBS> hi);
+  CtrlDamper(std::array<uint16_t, N_KNOBS> lo, std::array<uint16_t, N_KNOBS> hi);
   static constexpr uint16_t SKIP = 0xffff;
   CtrlEvent get();
   bool append(CtrlEvent event);
 
 private:
   uint8_t active = N_KNOBS;
-  std::array<uint8_t, N_KNOBS> thresh_lo;
-  std::array<uint8_t, N_KNOBS> thresh_hi;
+  std::array<uint16_t, N_KNOBS> thresh_lo;
+  std::array<uint16_t, N_KNOBS> thresh_hi;
   CtrlEvent latest = CtrlEvent(CtrlEvent::Switch, 0, 0);  // anything
   MovingAverage<uint32_t, 5> average[N_WHEN][N_KNOBS];
